wip 1

+++ database/migrations/1742203766424_create_access_tokens_table.ts
@@ -4,27 +4,31 @@ export default class extends BaseSchema {
   protected tableName = 'auth_access_tokens'
 
   async up() {
-    this.schema.createTable(this.tableName, (table) => {
-      table.increments('id')
-      table
-        .integer('tokenable_id')
-        .notNullable()
-        .unsigned()
-        .references('id')
-        .inTable('users')
-        .onDelete('CASCADE')
+    const hasTable = await this.schema.hasTable(this.tableName)
 
-      table.string('type').notNullable()
-      table.string('name').nullable()
-      table.string('hash').notNullable()
-      table.text('abilities').notNullable()
-      table.timestamp('created_at')
-      table.timestamp('updated_at')
-      table.timestamp('last_used_at').nullable()
-      table.timestamp('expires_at').nullable()
-    })
+    if (!hasTable) {
+      // Create table only if it doesn't exist
+      this.schema.createTable(this.tableName, (table) => {
+        table.increments('id')
+        table
+          .integer('tokenable_id')
+          .notNullable()
+          .unsigned()
+          .references('id')
+          .inTable('users')
+          .onDelete('CASCADE')
+
+        table.string('type').notNullable()
+        table.string('name').nullable()
+        table.string('hash').notNullable()
+        table.text('abilities').notNullable()
+        table.timestamp('created_at')
+        table.timestamp('updated_at')
+        table.timestamp('last_used_at').nullable()
+        table.timestamp('expires_at').nullable()
+      })
+    }
   }
-
   async down() {
     this.schema.dropTable(this.tableName)
   }
